+++ examples/interaction/combo/demo/cCircle.js
@@ -126,7 +126,7 @@ const graph = new G6.Graph({
     },
   },
   modes: {
-    default: ['drag-combo', 'drag-node', 'drag-canvas'],
+    default: ['drag-combo', 'drag-node', 'drag-canvas', 'click-select'],
   },
 });
 graph.data(data);
@@ -158,4 +158,15 @@ graph.on('combo:dragenter', (e) => {
 });
 graph.on('combo:dragleave', (e) => {
   graph.setItemState(e.item, 'dragenter', false);
+});
+
+
+graph.on('combo:mouseenter', (evt) => {
+  const { item } = evt;
+  graph.setItemState(item, 'active', true);
+});
+
+graph.on('combo:mouseleave', (evt) => {
+  const { item } = evt;
+  graph.setItemState(item, 'active', false);
 });