+++ examples/item/defaultEdges/demo/cubic2.js
@@ -62,18 +62,46 @@ const graph = new G6.Graph({
   modes: {
     default: ['drag-canvas'],
   },
-  defaultNode: {
-    style: {
-      fill: '#DEE9FF',
-      stroke: '#5B8FF9',
-    },
-  },
   defaultEdge: {
     type: 'cubic-horizontal',
-    style: {
-      stroke: '#F6BD16',
-    },
+    /* you can configure the global edge style as following lines */
+    // style: {
+    //   stroke: '#F6BD16',
+    // },
   },
+  /* styles for different states, there are built-in styles for states: active, inactive, selected, highlight, disable */
+  // edgeStateStyles: {
+  //   // edge style of active state
+  //   active: {
+  //     opacity: 0.5,
+  //     stroke: '#f00'
+  //   },
+  //   // edge style of selected state
+  //   selected: {
+  //     stroke: '#ff0'
+  //     lineWidth: 3,
+  //   },
+  // },
 });
 graph.data(data);
 graph.render();
+
+graph.on('edge:mouseenter', (evt) => {
+  const { item } = evt;
+  graph.setItemState(item, 'active', true);
+});
+
+graph.on('edge:mouseleave', (evt) => {
+  const { item } = evt;
+  graph.setItemState(item, 'active', false);
+});
+
+graph.on('edge:click', (evt) => {
+  const { item } = evt;
+  graph.setItemState(item, 'selected', true);
+});
+graph.on('canvas:click', (evt) => {
+  graph.getEdges().forEach(edge => {
+    graph.clearItemStates(edge);
+  });
+});