+++ examples/net/radialLayout/demo/sortRadial.js
@@ -438,8 +438,6 @@ const graph = new G6.Graph({
   },
   animate: true,
   defaultEdge: {
-    size: 1,
-    color: '#e2e2e2',
     style: {
       endArrow: {
         path: 'M 0,0 L 8,4 L 8,-4 Z',