+++ src/shape/edges/polyline.ts
@@ -22,7 +22,7 @@ Shape.registerEdge(
         x: 0,
         y: 0,
         stroke: Global.defaultEdge.style.stroke,
-        lineWidth: Global.defaultEdge.style.lineWidth
+        lineAppendWidth: Global.defaultEdge.style.lineAppendWidth
       },
       // 文本样式配置
       labelCfg: {